chore: remove unnecessary components and files.
