+++ src/components/PopupState.tsx
@@ -4,7 +4,6 @@ import { usePopupStore } from "../stores/popup";
 import AddUser from "../app/task/[type]/components/user/AddUser";
 import AssignTask from "@/app/task/[type]/components/task/AssignTask";
 import AddNewCustomer from "../app/customer/AddCustomer";
-import EditCustomer from "../app/customer/EditCustomer";
 import UpdateTask from "@/app/task/[type]/components/task/UpdateTask";
 import { UpdateAppointment } from "@/app/task/[type]/components/appointment/UpdateAppointment";
 
@@ -16,7 +15,6 @@ export default function PopupState() {
   if (popup === "ADD_USER") return <AddUser />;
   if (popup === "ASSIGN_TASK") return <AssignTask />;
   if (popup === "ADD_NEW_CUSTOMER") return <AddNewCustomer />;
-  if (popup === "EDIT_CUSTOMER") return <EditCustomer />;
 
   return null;
 }
