--- src/types/client.ts
@@ -1,11 +1,0 @@
-export interface ClientSource {
-  id: number;
-  source: string;
-}
-
-export interface ClientTag {
-  id: number;
-  tag: string;
-  textColor: string;
-  bgColor: string;
-}
